requirements.txt
activity==0.1
aiohttp==3.9.5
aiohttp-cors==0.8.1
aiosignal==1.3.2
aiounittest==1.3.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-common==1.1.28
azure-core==1.34.0
azure-identity==1.23.0
azure-mgmt-cognitiveservices==13.6.0
azure-mgmt-core==1.5.0
azure-search==1.0.0b2
azure-search-documents==11.5.3
Babel==2.9.1
botbuilder-core==4.17.0
botbuilder-dialogs==4.17.0
botbuilder-schema==4.17.0
botframework-connector==4.17.0
botframework-streaming==4.17.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
datedelta==1.4
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
emoji==1.7.0
fastapi==0.115.13
frozenlist==1.6.2
git-filter-repo==2.47.0
grapheme==0.6.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
jsonpickle==1.4.2
MarkupSafe==3.0.2
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.4.4
multipledispatch==1.0.0
oauthlib==3.2.2
openai==1.91.0
packaging==25.0
propcache==0.3.1
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
PyPDF2==3.0.1
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
yarl==1.20.0
